requirements.txt
PyQt5>=5.15.9
requests>=2.31.0
aiohttp>=3.9.0
PyInstaller>=6.10.0
python-docx>=1.1.0
markdown>=3.5.0
xhtml2pdf>=0.2.13

